Refactoring: Use abstract base class for EEP A4-04.
