+++ src/main/java/eu/aleon/aleoncean/device/remote/RemoteDeviceEEPA50401.java
@@ -15,109 +15,27 @@
  */
 package eu.aleon.aleoncean.device.remote;
 
-import java.util.Set;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import eu.aleon.aleoncean.device.DeviceParameter;
-import eu.aleon.aleoncean.device.DeviceParameterUpdatedInitiation;
-import eu.aleon.aleoncean.device.IllegalDeviceParameterException;
-import eu.aleon.aleoncean.device.RemoteDevice;
-import eu.aleon.aleoncean.device.StandardDevice;
 import eu.aleon.aleoncean.packet.EnOceanId;
-import eu.aleon.aleoncean.packet.RadioPacket;
 import eu.aleon.aleoncean.packet.radio.RadioPacket4BS;
+import eu.aleon.aleoncean.packet.radio.userdata.UserDataEEPA504;
 import eu.aleon.aleoncean.packet.radio.userdata.UserDataEEPA50401;
-import eu.aleon.aleoncean.packet.radio.userdata.UserDataScaleValueException;
 import eu.aleon.aleoncean.rxtx.ESP3Connector;
 
 /**
- *
- * @author Markus Rathgeb <maggu2810@gmail.com>
+ * Implementation for EPP A5-04-01.
+ * @author Stephan Meyer <smeyersdev@gmail.com>
  */
-public class RemoteDeviceEEPA50401 extends StandardDevice implements RemoteDevice {
+public class RemoteDeviceEEPA50401 extends RemoteDeviceEEPA504 {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(RemoteDeviceEEPA50401.class);
+	public RemoteDeviceEEPA50401(final ESP3Connector conn,
+								final EnOceanId addressRemote,
+								final EnOceanId addressLocal) {
+		super(conn, addressRemote, addressLocal);
+	}
 
-    private Double humidity;
-    private Double temperature;
-
-    public RemoteDeviceEEPA50401(final ESP3Connector conn,
-                                 final EnOceanId addressRemote,
-                                 final EnOceanId addressLocal) {
-        super(conn, addressRemote, addressLocal);
-    }
-
-    public Double getHumidity() {
-        return humidity;
-    }
-
-    public void setHumidity(final DeviceParameterUpdatedInitiation initiation, final Double humidity) {
-        final Double oldHumidity = this.humidity;
-        this.humidity = humidity;
-        fireParameterChanged(DeviceParameter.HUMIDITY_PERCENT, initiation, oldHumidity, humidity);
-    }
-
-    public Double getTemperature() {
-        return temperature;
-    }
-
-    public void setTemperature(final DeviceParameterUpdatedInitiation initiation, final Double temperature) {
-        final Double oldTemperature = this.temperature;
-        this.temperature = temperature;
-        fireParameterChanged(DeviceParameter.TEMPERATURE_CELSIUS, initiation, oldTemperature, temperature);
-    }
-
-    private void parseRadioPacket4BS(final RadioPacket4BS packet) {
-        if (packet.isTeachIn()) {
-            LOGGER.debug("Ignore teach-in packets.");
-            return;
-        }
-
-        final UserDataEEPA50401 userData = new UserDataEEPA50401(packet.getUserDataRaw());
-        try {
-            setHumidity(DeviceParameterUpdatedInitiation.RADIO_PACKET, userData.getHumidity());
-        } catch (final UserDataScaleValueException ex) {
-            LOGGER.warn("Received humidity is invalid.");
-        }
-
-        try {
-            setTemperature(DeviceParameterUpdatedInitiation.RADIO_PACKET, userData.getTemperature());
-        } catch (final UserDataScaleValueException ex) {
-            LOGGER.warn("Received temperature is invalid.");
-        }
-    }
-
-    @Override
-    public void parseRadioPacket(final RadioPacket packet) {
-        if (packet instanceof RadioPacket4BS) {
-            parseRadioPacket4BS((RadioPacket4BS) packet);
-        } else {
-            LOGGER.warn("Don't know how to handle radio choice 0x%02X.", packet.getChoice());
-        }
-    }
-
-    @Override
-    protected void fillParameters(final Set<DeviceParameter> params) {
-        params.add(DeviceParameter.HUMIDITY_PERCENT);
-        params.add(DeviceParameter.TEMPERATURE_CELSIUS);
-    }
-
-    @Override
-    public Object getByParameter(final DeviceParameter parameter) throws IllegalDeviceParameterException {
-        switch (parameter) {
-            case HUMIDITY_PERCENT:
-                return getHumidity();
-            case TEMPERATURE_CELSIUS:
-                return getTemperature();
-            default:
-                return super.getByParameter(parameter);
-        }
-    }
-
-    @Override
-    public void setByParameter(final DeviceParameter parameter, final Object value) throws IllegalDeviceParameterException {
-        assert DeviceParameter.getSupportedClass(parameter).isAssignableFrom(value.getClass());
-        super.setByParameter(parameter, value);
-    }
+	@Override
+	protected UserDataEEPA504 createUserData(RadioPacket4BS packet) {
+		return  new UserDataEEPA50401(packet.getUserDataRaw());
+	}
 
 }
